Integrate and some little refactoring
