--- src/VSCodeExtension/src/formatter-core.ts
@@ -1,9 +1,0 @@
-export type FormatRule = (code: string) => string;
-
-export const formatter = (code: string, rules: FormatRule[]): string => {
-    return rules.reduce((formattedCode, rule) => rule(formattedCode), code);
-}
-
-export const spaceAfterIf: FormatRule = (code: string): string => {
-    return code.replace(/(^| +|;|})if[ \n]*/, "$1if ");
-}
