--- src/VSCodeExtension/src/formatter/index.d.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export type FormatRule = (code: string) => string;
