+++ src/VSCodeExtension/src/formatter/rules/space-after-if.ts
@@ -1,0 +1,5 @@
+import { FormatRule } from "../formatter";
+
+export const spaceAfterIf: FormatRule = (code: string): string => {
+    return code.replace(/(^| +|;|})if[ \n]*/, "$1if ");
+}
